Started migration to twitter4j 2.2.1. Code is compileable, but auth is not working for now. Changed id types from int to long.

+++ YFrogAndroid/src/com/codeminders/yfrog2/android/model/TwitterDirectMessage.java
@@ -13,7 +13,7 @@ import java.util.Date;
  *
  */
 public class TwitterDirectMessage implements Serializable {
-	private int id;
+	private long id;
 	private TwitterUser sender;
 	private String text;
 	private Date createdAt;
@@ -21,13 +21,13 @@ public class TwitterDirectMessage implements Serializable {
 	/**
 	 * @return the id
 	 */
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	/**
@@ -68,7 +68,7 @@ public class TwitterDirectMessage implements Serializable {
 	}
 	
 	private void writeObject(ObjectOutputStream out) throws IOException {
-		out.writeInt(id);
+		out.writeLong(id);
 		out.writeObject(sender);
 		out.writeUTF(text);
 		out.writeObject(createdAt);
